Branch daniel journal

+++ Project2Solution/Project.Client/Controllers/SongController.cs
@@ -48,16 +48,26 @@ namespace Project.Client.Controllers
         [Produces("application/json")]
         public async Task<Song> Get([FromRoute]int id)
         {
-            Song song = await Task.Run(() => Mapper.Map(repository.GetSongById(id)));
-
-            return song;
+            try
+            {
+                Song song = await Task.Run(() => Mapper.Map(repository.GetSongById(id)));
+                return song;
+            }
+            catch
+            {
+                return new Song();
+            }
         }
 
         // POST api/values
         [EnableCors]// it will use default policy
         [HttpPost]
+        [ProducesResponseType(typeof(Song), StatusCodes.Status201Created)]
+        [Produces("application/json")]
         public async Task<IActionResult> Post([FromBody] Song song)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             if (song != null)
             {
                 int rowAffected = await Task.Run(() => repository.CreateSong(Mapper.Map(song)));
@@ -71,8 +81,12 @@ namespace Project.Client.Controllers
 
         //PUT api/values/5
         [HttpPut]
+        [ProducesResponseType(typeof(Song), StatusCodes.Status202Accepted)]
+        [Produces("application/json")]
         public async Task<IActionResult> Put([FromBody] Song song)
         {
+            if (!ModelState.IsValid) return BadRequest();
+
             int rowAffected = await Task.Run(() =>
             repository.UpdateSong(Mapper.Map(song)));
 
@@ -82,13 +96,24 @@ namespace Project.Client.Controllers
 
         // DELETE api/values/5
         [HttpDelete("{id}")]
+        [ProducesResponseType(typeof(Song), StatusCodes.Status200OK)]
+        [Produces("application/json")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
-            int rowAffected = await Task.Run(() => repository.DeleteSong(id));
+            try
+            {
+                if (!ModelState.IsValid) return BadRequest();
 
-            if (rowAffected > 0) return Ok();
+                int rowAffected = await Task.Run(() => repository.DeleteSong(id));
 
-            return BadRequest();
+                if (rowAffected > 0) return Ok();
+
+                return BadRequest();
+            }
+            catch
+            {
+                return NoContent();
+            }
         }
     }
 }
